requirements.txt
torch>=2.1.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scipy>=1.11.0
tqdm>=4.65.0
pyyaml>=6.0
gurobipy>=11.0
holidays>=0.46
